Dont install deps in mypy CI
Originally we were using mypy plugins and had to have those installed to run it. Now we are not, so we can get away with just installing mypy, to save in some CI minutes.

Once we add mypy plugins back, we can add just the ones we need to this install.