First pass at documenting internals
This closes https://github.com/LineaLabs/lineapy/issues/410 by creating an INTERNALS document.

As I was writing this, I noticed that it would have been better to iteratively document these things as we progressed. One of my goals in the future then as I change features or add new ones is to make sure they are added here.

Also, this document is very large and has some links between it. One way we could improve it is by using Sphinx and Juptyer Book to build documentation. That way this could have a TOC or be split into multiple files to be browsed more easily. We could also then use the `autodoc` feature to pull in docstrings of modules and function, and refer to them directly, instead of linking to the source code they are defined in.

Feel free to edit as you like and ping me via email with any questions.