Move side effect generation into context
This change is in preperation for creating side effects based on sys.settrace.
It doesn't change any behavior.

The change is neccesary so that we can add a `side_effects` attribute to the
context. Then, in the `l_exec` we can add to this `side_effects` part of the
context, after using `settrace` to understand what side effects to add.