Replace artifact version validator with artifact version parser
Instead of returning whether or not the version of the artifact
the user passes in is valid, parse it into a valid version
and throw an exception if the parsing fails. This simplifies
the code a lot as you can assume the artifact version is
a positive int or either 'all' or 'latest'.