Log successful delete
Inform the user of the name and version of the artifact they deleted.
Remove check that the number of artifacts pointing to a pickle is one
because a new pickle is created everytime an artifact is saved.
Remove extra test because it is already being tested in
test_delete_version_complex.