LIN-441 Add artifact version validation to Lineapy API (#685)
* restructure to reduce nesting

* Reformat with black

(cherry picked from commit 693a818c6b94d235c2f3a912d175bd90eaf0a479)

* remove wildcards - will be moot later

* validate version before deleting artifact

* update unit tests

* Test is_artifact_version_valid

Write a unit test to test is_artifact_version_valid.
Reuse is_artifact_version_valid in db delete_artifact_by_name.

(cherry picked from commit ed47b731cda1151a120a91529868074db933a3dc)

* Replace artifact version validator with artifact version parser

Instead of returning whether or not the version of the artifact
the user passes in is valid, parse it into a valid version
and throw an exception if the parsing fails. This simplifies
the code a lot as you can assume the artifact version is
a positive int or either 'all' or 'latest'.

* Fix mypy issues

Use different variable names for different types in
parse_artifact_version.

* Log successful delete

Inform the user of the name and version of the artifact they deleted.
Remove check that the number of artifacts pointing to a pickle is one
because a new pickle is created everytime an artifact is saved.
Remove extra test because it is already being tested in
test_delete_version_complex.

* Log when value node not found in DB

* Add ARTIFACT_MIN_VERSION constant instead of hardcoding zero

We might start at version 1 in the future so creating a constant is a
good idea.
Add test for lineapy.get.
Remove int casting case in parse_artifact_version because float casting
handles it.

* Remove is_artifact_version_valid and test_get_version_zero snapshot

is_artifact_verion_valid is unneeded since the introduction of
parse_artifact_version.
Since test_get_version uses asserts, snapshots are not needed.

* Validate lineapy.get version input

Pass version through parse_artifact_version to ensure user
passes in a valid version.

* Fix mypy issues in get_artifact_by_name

Perform type acrobatics to convert version types accepted by
get_artifact_by_name (Optional[int]) to types accepted by
parse_artifact_by_name
(Union[int, str]).

* Reformat with black

* Remove commented out code and add raises to docstring

Accidentally left commented code in get_artifact_by_name.
Add raises ValueError to docstring of delete API call.

* Change pickle not found logs in delete to debug

If no pickle was found to delete, the user does not need to know.

* Rewrite parse_artifact version to remove extaneous casts

Check to see if the type of version is string or int first and
handle those cases. If that fails, try casting.
Since the user will be passing ints or strings most of the time
casting is a waste of resources if not needed.

* Shorten version conversion in lineally.get with ternary

The function parse_artifact_version does not take None, but lineapy.get does. 
If None it passed in it needs to be converted into 'latest'.
We can use a ternary operator to do that, which will shorten the code.

Co-authored-by: shardul <shardul@linea.ai>

* Use python syntax in ternary operator

Java syntax was accidentally used in ternary operator.

* Reformat with black

* Add asserts to parse_artifact_version_test

Before the test was just checking to make sure the function raised
exceptions when passed invalid input.
Now it also makes sure the parsed output when passed valid input is
correct.

* Reuse error messages in parse_artifact_version

Personalize the first part of the error message to reveal the reason the
parse failed. Standardize the second part of it to tell the user what is
accepted as a valid version.
If version is passed in as a string, convert to lower case.

Co-authored-by: shardul <shardul@linea.ai>