Create temp linea_pickle fold in pytest fixture after folder move
If no linea_pickle folder exists, lineapy will issue a warning before
creating the folder. This warning causes some tests in test_script to fail as they do output matching. Since the creation of the folder is expected, go ahead and to it in the fixture to avoid the warning.