LIN-442 Fix pickle IO leak (#707)
* Convert pickle folder pytest fixture to autouse

There is practically no performance hit to making the pytest fixture
autouse.
Delete the explicit move_folder fixture adds to avoid moving the folder
twice.

* Create temp linea_pickle fold in pytest fixture after folder move

If no linea_pickle folder exists, lineapy will issue a warning before
creating the folder. This warning causes some tests in test_script to fail as they do output matching. Since the creation of the folder is expected, go ahead and to it in the fixture to avoid the warning.

* Reformat code with black

* Fix utils tests

The tests in test_config were changing the artifact_storage_dir and not
resetting it after finishing up the test. This was messing up the
move_artifact_dir fixture.

* Parametrize remove_file_ext test

Instead of looping through cases in a single test, use
pytest.mark.parametrize to handle multiple cases.

* Remove extraneous move_folder fixture from pickle writing test

Since moving the linea_pickles folder happens automatically for every
test, specifying the fixture causes it to happen twice.
This causes problems when moving the folder back.

* Add unused pytest import

* Delete accidental to_pipeline result files