Fetch artifact list from db but not dependencies
It is assumed that all dependencies will be artifacts in the artifact
list so dependencies do not need to be fetched separately. Instead they
can be reused when fetching a Pipeline.

Replace some object strings in sqlalchmy relationships with the ORM
object types themselves
Mypy was complaining about being unable to pick up on the type of a
relationship.