Lin 447 pipeline object (#710)
* Add Pipeline and PippelineORM stub

Create PipelineORM with many to many relationship with artifacts.
Create Pipeline API class containing a TaskGraph.

* Add generate_pipeline function stub in Pipeline object

Fix broken Table import in relational
Add dependency as TaskGraph member so it can be used in pipeline
generation.

* Rename pipeline artifact relationship table string key

The string key was just association so it was renamed to
artifact_to_pipeline.

* Implement generate_pipeline in Pipeline object

Copy logic from to_pipeline function and modify the code to fit Pipeline
input parameters.

* Reformat with black

* Pass TaskGraph into plugins directly

Since Pipeline is an object now, the TaskGraph can be initialized when
the object is instead of when the orchestration occurs.

* Re add to_pipeline functionality

Have to_pipeline create a Pipeline object and call generate_pipeline and
return the result.

* add todos

* revert unintentional rename

* move the api_utils and add task_graph creator as a util

* add create_pipeline and clean up lint issues. make a placeholder for saving pipelines.

* fix cli, add a test case to test airflow export and task dependencies passed via cli

* Start Working on Pipeline Save

Add IDs to Artifact and Pipeline

* fixes to save pipelineorm with proper autoincrementing ids.

* add a few ignores for pytest and fix issues with tests.

* Add get_pipeline_by_name to db

Allow user to get PipelineORM object from database given name

* Implement get_pipeline API call

Return a Pipeline object by name.
Remove default value for PipelineORM name column because it is
nonnullable so no default is needed.

* Add simple test for pipeline

* Add ArtifactDependency ORM

This represents the one to many relationship between an artifact and
it's sources in the context of a pipeline.

* Add ArtifactDependencyORM relationships

* Handle dependencies when writing pipelines

* Fix tests by specifying foreign_keys in relationships

* Add more Pipeline object tests

* Add version as unique constraint

Makes copy notebokk tests pass.
Add currently failing pytest dependency test.

* Use relationship table for pre artifacts in artifact dependency

In order to construct a proper many to many relationship between pre
artifacts and dependencies, a secondary table must be constructed in
sqlalchemy.

Convert relationship between Pipeline and dependencies into a
bidirectional relationship.

* Fetch dependencies when fetching Pipeline from db

The dependencies must be fetched from the database and converted
into a TaskGraphEdge to pass into Pipeline before returning the object.

* Fetch artifact list from db but not dependencies

It is assumed that all dependencies will be artifacts in the artifact
list so dependencies do not need to be fetched separately. Instead they
can be reused when fetching a Pipeline.

Replace some object strings in sqlalchmy relationships with the ORM
object types themselves
Mypy was complaining about being unable to pick up on the type of a
relationship.

* Skip fetching dependency for Pipeline if post artifact is None

If post artifact is None, the code would crash because none has no
attribute name. Instead we want to skip attempting to fetch an
incomplete dependency.

* Add complex pipeline dependency test

Test creating and fetching two pipelines with overlapping dependencies.

* Add test to check pre artifact overlap in pipeline dependency

Make sure the db can handle multiple dependencies with the same pipeline
id and pre artifacts.

Disallow creation of empty pipelines.

* Simplify pipeline tests

Reduce number of tests to 3. The first test tests multiple pipelines
with different artifact lists. The second test tests multiple pipelines
with overlapping dependencies. The third test tests a signle pipeline
with overlapping dependencies.

Co-authored-by: Shardul Sardesai <shardul@linea.ai>