LIN-737: Fix argo tests requiring /home/runner (#865)
# Description

Argo tests at this time snapshot the user home directory, which means the snapshot tests must be manually updated to pass on github actions.

This PR moves the path expansion to the dag file itself, meaning the generated DAG will not depend on the users home directory. 

Also added and fixed hints for how to use the Dockerfile for argo. 

Fixes LIN-737

## Type of change

Please delete options that are not relevant.

- [X] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?
Tested locally with housing example, both providing and not providing a kube config
`        "kube_config": "/Users/andrewcui/.kube/config",`